Initial Project Setup
This is an initialization of the Octave Player netbeans project for the CSI3370 project. The octave_player package contains skeletons of all of the classes per the design doc w/the exception of the AudioStream class, which will simply be a instance of the javafx mediaPlayer class as a data member of the driver/main class.