Main method reads finds playlist files, added observer/observable interfaces
The main method now creates a Playlist list, Queue, OctaveView, and OctaveController. IT creates the playlist list by fidning all the files with ".opl" extensions in the runtime directory and creating a playlist object for each of those playlist files.

Also created observer and obsevable interfaces. View is an observer and it binds observers which will tell View when they have changes. Playlists and the queue are observables and will bind a view to themselves when they are created. This allows us to implement the MVC architecture with the Model classes as observables,a nd teh view class as an observer.