Cleaned up Main method, fixed view/controller constructor headers
Per the MVC design with the observer/observable interfaces, the OvtaveView constructor is passed primary stage and controller parameters. This allows the view  to have access to the stage, and provides it a controller object for modifying the models.

Similarly controller now has access to the driver class. This allows the controller to access the model classes via getter methods in the driver class.  This makes the controller the means via which the view can edit the playlist or queue on various user-triggered events.