Added 1st draft of deletePlaylist method
deltePlaylist method is implemented in driver class. Once view comes online, it will be extended to alert users if the playlist to be deleted does not exist.  Added a simple test method to the controller to test this method. Deletes the playlist created by playlistCreation method in the controller used for similar testing. Also set the view's current data members to private.